fix(ros-testing)!: require `runs-on` input to workflow call
Depending on the `ros-distro`, the correct `runs-on` OS needs to be used, because the ROS distro might not be found. E.g., `foxy` cannot be found on `ubuntu-22.04`, but can be found on `ubuntu-20.04`. The fix delegates the correct choice to the callee.

BREAKING CHANGE: Callees need to call the workflow with `runs-on`.